+++ CHANGELOG.md
@@ -1,7 +1,16 @@
-# v0.1.1 - Snapshot
+## v0.2.0 - Velocity Support, Updates & Fixes
+* Added sonatype
+* Added gradle kotlin dsl
+* Added velocity support
+* Fixed issue where cancelled bungee tasks won't start again
+* Updated workflows
+* Updated dependencies
+* Updated README.md
+
+## v0.1.1 - Snapshot
 * Migration to JitPack
 * Added JitCI badges
 * Fixed description
 
-# v0.1.0 - Snapshot
+## v0.1.0 - Snapshot
 Hello, World!